requirements.txt
discord.py==2.3.2
aiohttp==3.9.3
python-dotenv==1.0.1
audioop-lts; python_version>='3.13' 